backend/requirements.txt
fastapi
uvicorn[standard]
psutil
requests
httpx
sqlmodel
ldap3
pyjwt
bcrypt
websockets